app/requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
azure-cosmos==4.5.1
python-dotenv==1.0.0
uvicorn==0.24.0
aiohttp==3.9.1
azure-core==1.30.2
azure-identity==1.15.0
PyJWT==2.8.0
cryptography==41.0.7
httpx==0.25.2
python-jose[cryptography]==3.5.0
pytest==8.4.1
azure-storage-blob==12.25.1
